Allow all networking interfaces to be defined as ports
Rather than require users to use the networks and subnets to define a
networking interface for an instance, users should be able to just
define them using the ports api if they choose to. This removes
validation checking that a user entered at least one network and subnet,
and insteand checks that a user entered at least one network, subnet, or
port.