Bug 1955969: Make port names unique
The bugzilla in question is rooted on attempting to create
ports with the same name.

This patch uses the subnet and netids as name suffixes for ports.

Also limits the length of the port name to no more than 255 characters,
which is in line with what the current neutron db schema can tolerate.