Remove legacy cruft from Makefile
Makefile contained a build-cross target which appears to have been for
cross-compiling for a number of different targets. However, it was
building 'cmd/openstack-machine-controller', which does not exist, so it
is clearly not in use.

build-cross is removed, along with everything which depends on it, and
everything which is only used by removed targets.