Refactor MAPO to use a reconcile model
Primarily, both Create() and Update() become simple wrappers around
reconcile(). This allows us to concentrate on ensuring reconcile() is
idempotent, and gives us greater scope for retrying on error.

As a result we are able to remove many cases which put a machine in the
failed state. This includes any error on creation which doesn't relate
to an invalid state or a failure in the actual create call. With this
update we now never set the failed state in the actuator on update.