Set controller's SyncPeriod to 1 hour
The Machine controller that we run doesn't set the SyncPeriod. This means that
Machine objects are not periodically reconciled with the status of them in
OpenStack.

This is a problem when the user adds a floating IP to a Server manually. In
such case cloud-provider-openstack will notice it and update the Node object
accordingly. This will cause Kubelet to generate a new CSR with updated SANs
including new FIP address. Then cluster-machine-approver will compare the list
of SANs from the CSR with Machine's Status.Addresses and will never find the
FIP failing the approval.

This commit sets SyncPeriod of the controller to 1 hour in order to make
sure Machines are periodically reconciled with the state of the
instances in OpenStack just like Nodes are.