Don't build InstanceSpec during delete operations
We were building a full InstanceSpec struct as part of our Delete
operation before promptly throwing the bulk of it away as we only needed
the root volume information.

This is wasteful as building the InstanceSpec involves looking up
resources to resolve both identifiers (like 'serverGroupName', which
must be resolved to an ID of a new or existing server group using the
Compute API) and filters (such as a 'networks[*].filter.subnets[*]',
which must be resolved to an ID if a network name is given).

It's also the source of bugs. This is because we currently do not
validate or propagate changes to a 'Machine' after initial creation
meaning a user can update their 'Machine' to use e.g. an invalid network
name. While this would have no immediate effect - remember, the changes
aren't propagated - it would prevent the users deleting the Machine
since validation would fail repeatedly.

Avoid all of this by not building the whole InstanceSpec. Instead, build
only what we need: the root volume information.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>