Rename variables
The 'MachineSpecFromProviderSpec' function returns an OpenStack-specific
ProviderSpec struct. Rename the variables in an attempt to clarify this.

Conflicts:
  pkg/machine/actuator.go

NOTE(stephenfin): Conflicts are due to the v1alpha7 API bump present on
master. In this case, there are a number of functions whose signature
has changed and we need to revert to the older signature.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
(cherry picked from commit 9e835b72728e6bb06f72483598532fbcf8b5d731)