Add tests for MachineToInstanceSpec
These are very preliminary but it's something to build on going forward.

Conflicts:
  pkg/machine/convert.go
  pkg/machine/convert_test.go

NOTE(stephenfin): Conflicts are due to the use of CAPO API v1alpha6
here, rather than v1alpha7 used in 4.15. The two main differences in
v1alpha7 that affect us here are the replacement of SecurityGroupParam
with SecurityGroupFilter and the additional of AdditionalBlockDevices.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
(cherry picked from commit 0bdf38346086721598df793bf6521432c51488d2)