Revendor CAPO
Point to the `release-4.15` branch of downstream CAPO, that is tracking
upstream's `release-0.8` branch.

Because the `v0.8.0` tag upstream points to a commit that does not
belong to the `release-0.8` branch, we're getting a `v0.8.0-beta-xxx`
version in the go mod but this is indeed the stable v0.8 release branch.