OCPBUGS-33806: Fix panic deleting Machine in ERROR state
CAPO v0.9 needs to resolve port names when deleting a server in ERROR
state. This would panic because we were not passing in a port list, so
CAPO looked at the cluster network, which we were also not passing in.
Additionally the port names need the instance name to be passed in in
order to be resolve correctly.

This change also deletes a bunch of dead code. Deleting it highlighted
that:
* MachineToInstanceSpec does not use its networkService parameter
* networkParamToCapov1PortOpt cannot actually return an error

We update these signatures and the tests accordingly.